# backend/docs/openapi-gpt-action.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Isekai Lucas Engine GPT Actions",
    "version": "0.22.0",
    "description": "Schema normal para GPT Actions. Expone endpoints de lectura, busqueda, preview y mutacion central controlada. No expone rollback, admin ni mutadores peligrosos de combate/misiones."
  },
  "servers": [
    {
      "url": "https://isekai-lucas-engine.onrender.com"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
      "GenericResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "additionalProperties": true
      },
...
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/api/search/docs": {
      "get": {
        "operationId": "searchDocs",
        "summary": "Busca chunks indexados de rules_engine.md y world_bible.md.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resultados de busqueda documental.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/npcs/{npcId}/full": {
      "get": {
        "operationId": "getNpcFull",
        "summary": "Obtiene ficha completa de un NPC.",
        "parameters": [
          {
            "name": "npcId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "NPC completo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/locations/{locationId}/full": {
      "get": {
        "operationId": "getLocationFull",
        "summary": "Obtiene ficha completa de una ubicacion.",
        "parameters": [
          {
            "name": "locationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ubicacion completa.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}